fix(time): return f64 from rt_time_now_seconds instead of truncating to i64
Was casting f64→i64 losing all sub-second precision. GPU benchmarks now
report accurate microsecond-level frame times (4.07ms/frame, 510 Mpix/s).

+++ src/compiler_rust/compiler/src/interpreter_extern/time.rs
@@ -17,7 +17,7 @@ use crate::value::Value;
 pub fn rt_time_now_seconds(_args: &[Value]) -> Result<Value, CompileError> {
     unsafe {
         let time = simple_runtime::value::rt_time_now_seconds();
-        Ok(Value::Int(time as i64))
+        Ok(Value::Float(time))
     }
 }
